Implemented routing and controller
